separate gaudi components from lib
